Renamed and restructured several components Changed Tile to be a header component Changed from departuretile to busstoptile Changed classnames to reflect change
