--- src/dashboards/BusStop/DepartureTile/DepartureTile.scss
@@ -1,39 +1,0 @@
-@use '../BusStopDashboard.scss';
-
-.bus-stop-departure-tile {
-    &-head {
-        border-bottom: none;
-    }
-
-    &-head-icon {
-        width: 10%;
-    }
-
-    &-head-departure {
-        width: 18%;
-    }
-
-    &-head-track {
-        width: 18%;
-    }
-
-    &-head-situation {
-        width: 30%;
-    }
-}
-
-@media (max-width: BusStopDashboard.$breakpoint-mobile) {
-    .bus-stop-departure-tile {
-        &-head-departure {
-            width: 22%;
-        }
-
-        &-head-track {
-            width: 13%;
-        }
-
-        &-head-situation {
-            width: 11%;
-        }
-    }
-}
