+++ src/dashboards/BusStop/components/BusStopHeader/BusStopHeader.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Heading2 } from '@entur/typography'
 import { WalkInfo } from '../../../../logic/use-walk-info/useWalkInfo'
-import './Tile.scss'
+import './BusStopHeader.scss'
 
 function formatWalkInfo(walkInfo: WalkInfo) {
     if (walkInfo.walkTime / 60 < 1) {
@@ -13,9 +13,9 @@ function formatWalkInfo(walkInfo: WalkInfo) {
     }
 }
 
-function Tile({ title, icons, walkInfo, children }: Props): JSX.Element {
+function TileHeader({ title, icons, walkInfo }: Props): JSX.Element {
     return (
-        <div className="bus-stop-tile">
+        <>
             <header className="bus-stop-tile-header">
                 <Heading2 className="bus-stop-tile-header-heading">
                     {title}
@@ -23,12 +23,11 @@ function Tile({ title, icons, walkInfo, children }: Props): JSX.Element {
                 <div className="bus-stop-tile-header-icons">{icons}</div>
             </header>
             {walkInfo ? (
-                <div className="bus-stop-tile-walking-time">
+                <div className="bus-stop-tile-header-walking-time">
                     {formatWalkInfo(walkInfo)}
                 </div>
             ) : null}
-            {children}
-        </div>
+        </>
     )
 }
 
@@ -36,7 +35,6 @@ interface Props {
     title: string
     icons: JSX.Element | JSX.Element[]
     walkInfo?: WalkInfo
-    children: JSX.Element[] | JSX.Element
 }
 
-export { Tile }
+export { TileHeader }
