+++ src/dashboards/BusStop/components/TableRow/TableRow.scss
@@ -12,6 +12,9 @@
         line-height: 2.25rem;
         font-weight: 600;
         color: var(--primary-text-color);
+        overflow: hidden;
+        white-space: nowrap;
+        text-overflow: ellipsis;
     }
 
     &-time,
