--- src/dashboards/BusStop/components/Tile/Tile.scss
@@ -1,35 +1,0 @@
-.bus-stop-tile {
-    width: 100%;
-    max-height: 80vh;
-    padding: 2rem;
-    margin-bottom: 2rem;
-    overflow: scroll;
-    border-radius: 1rem;
-
-    background-color: var(--tavla-box-background-color);
-    color: var(--tavla-font-color);
-
-    &-header-icons {
-        font-size: 2rem;
-        display: flex;
-        align-items: center;
-    }
-
-    &-header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 1rem;
-
-        &-heading {
-            margin: 0;
-        }
-    }
-
-    &-walking-time {
-        font-style: normal;
-        font-weight: normal;
-        font-size: 1rem;
-        color: var(--tavla-font-color);
-    }
-}
